Move some binops Series tests
